[1366] Move RC helper methods
Into our generic recruitment cycle helper
