+++ app/helpers/recruitment_cycle_helper.rb
@@ -12,4 +12,24 @@ module RecruitmentCycleHelper
   def previous_recruitment_cycle_period_text
     "#{Settings.current_recruitment_cycle_year - 1} to #{Settings.current_recruitment_cycle_year}"
   end
+
+  def hint_text_for_mid_cycle
+    "#{I18n.t('find.cycles.today_is_mid_cycle.description')} (#{Find::CycleTimetable.find_opens.to_fs(:govuk_date)} to #{Find::CycleTimetable.apply_2_deadline.to_fs(:govuk_date)})"
+  end
+
+  def hint_text_for_after_apply_1_deadline_passed
+    "#{I18n.t('find.cycles.today_is_after_apply_1_deadline_passed.description')} (#{Find::CycleTimetable.apply_1_deadline.to_fs(:govuk_date)} to #{Find::CycleTimetable.apply_2_deadline.to_fs(:govuk_date)})"
+  end
+
+  def hint_text_for_after_apply_2_deadline_passed
+    "#{I18n.t('find.cycles.today_is_after_apply_2_deadline_passed.description')} (#{Find::CycleTimetable.apply_2_deadline.to_fs(:govuk_date)} to #{Find::CycleTimetable.find_closes.to_fs(:govuk_date)})"
+  end
+
+  def hint_text_for_today_is_after_find_closes
+    "#{I18n.t('find.cycles.today_is_after_find_closes.description')} (#{Find::CycleTimetable.find_closes.to_fs(:govuk_date)} to #{Find::CycleTimetable.find_reopens.to_fs(:govuk_date)})"
+  end
+
+  def hint_text_for_today_is_after_find_opens
+    "#{I18n.t('find.cycles.today_is_after_find_opens.description')} (#{Find::CycleTimetable.find_reopens.to_fs(:govuk_date)})"
+  end
 end
